Used Mapping to display carnage stats, added SamaHelper
