+++ src/SamaHelper.js
@@ -1,0 +1,13 @@
+import stone from './assets/img/mmaf/aStone.png';
+import wood from './assets/img/mmaf/aWood.png';
+import iron from './assets/img/mmaf/aIron.png';
+import gold from './assets/img/mmaf/aGold.png';
+import exp from './assets/img/mmaf/aExp.png';
+import grain from './assets/img/mmaf/aGrain.png';
+import string from './assets/img/mmaf/aString.png';
+import fish_specimen from './assets/img/mmaf/aFishSpecimen.png';
+
+export const ressourcesIcons = 
+{
+    alpha:{stone,wood,iron,gold ,exp ,grain,string,fish_specimen}
+}
